pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bwf_forms"
version = "0.1"
dependencies = [
    "django>=5.2",
    "python-decouple>=3.8",
    "dj-database-url",
    "djangorestframework",
    "git+https://github.com/dbca-wa/webtemplate2.git#egg=webtemplate_dbca",
]
description = "A Django app to create and render web forms."
readme = "README.rst"
requires-python = ">= 3.10"
authors = [
    {name = "Juan Achury", email = "juan.achurybueno@dbca.wa.gov.au"},
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
[project.scripts]
"manage.py" = "bwf_forms:django_manage"

[project.urls]
Homepage = "https://www.dbca.wa.gov.au/"
